+++ packages/config/eslint-config/react-library.js
@@ -98,6 +98,7 @@ module.exports = {
     "@typescript-eslint/no-unused-expressions": "off",
     "no-underscore-dangle": "off",
     "no-console": "off",
+    "no-void": "off",
     "no-nested-ternary": "off",
   },
 };