+++ packages/components/image-upload/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/image-upload
 
+## 0.0.31
+
+### Patch Changes
+
+- fix
+
 ## 0.0.30
 
 ### Patch Changes