+++ packages/config/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/eslint-config
 
+## 0.0.20
+
+### Patch Changes
+
+- fix
+
 ## 0.0.19
 
 ### Patch Changes